execution/requirements.txt
functions-framework==3.*
google-cloud-firestore==2.*
google-cloud-secret-manager==2.*
simple-salesforce==1.*
